refactor: step5 resultView에서 우승자를 찾는 비지니스 로직 제거
