+++ src/main/java/carracing/domain/race/Race.java
@@ -1,4 +1,4 @@
-package carracing.domain;
+package carracing.domain.race;
 
 import carracing.domain.car.Car;
 import carracing.domain.car.Name;
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Race {
-    private final int rounds;
+    private final int totalRoundNumber;
     private final RandomNumberGenerator randomNumberGenerator;
     private final List<RoundRecord> roundRecords = new ArrayList<>();
 
@@ -18,7 +18,7 @@ public class Race {
 
     private Race(List<Car> cars, int rounds, RandomNumberGenerator randomNumberGenerator) {
         this.randomNumberGenerator = randomNumberGenerator;
-        this.rounds = rounds;
+        this.totalRoundNumber = rounds;
         this.cars.addAll(cars);
     }
 
@@ -27,16 +27,10 @@ public class Race {
         return new Race(cars, rounds, randomNumberGenerator);
     }
 
-    private static List<Car> createCars(List<String> carNames) {
-        return carNames.stream()
-                .map(name -> Car.from(Name.from(name)))
-                .collect(Collectors.toUnmodifiableList());
-    }
-
     public List<RoundRecord> start() {
         recordRound();
 
-        for (int i = 0; i < rounds; i++) {
+        for (int i = 0; i < totalRoundNumber; i++) {
             moveCars();
             recordRound();
         }
@@ -44,19 +38,29 @@ public class Race {
         return roundRecords;
     }
 
+    public List<Car> getWinners() {
+        RoundRecord lastRound = roundRecords.get(roundRecords.size() - 1);
+        return lastRound.getLeadingCar();
+    }
+
+    public List<Car> getCars() {
+        return cars;
+    }
+
+    private static List<Car> createCars(List<String> carNames) {
+        return carNames.stream()
+                .map(name -> Car.from(Name.from(name)))
+                .collect(Collectors.toUnmodifiableList());
+    }
+
     private void recordRound() {
         roundRecords.add(RoundRecord.from(cars));
     }
-
     private void moveCars() {
         cars = cars.stream()
                 .map(car -> {
                     int randomDistance = randomNumberGenerator.generate();
                     return car.move(randomDistance);
                 }).collect(Collectors.toUnmodifiableList());
-    }
-
-    public List<Car> getCars() {
-        return cars;
     }
 }
